refactor(enc): address PR-A pre-push gate findings
Seven low-risk fixes confined to files this PR already added: use copy
(memmove-correct) instead of a hand-rolled reverse loop in
Filterbank.shiftIn; correct the FMA-barrier doc comments in
filterbank.go and mdct.go, which wrongly claimed the window/z-population
float64() wraps were load-bearing arm64 barriers (they feed a store, not
an adjacent +/-; the matrixStep and MDCTGranule inner-sum wraps are the
real barriers, verified via GOARCH=arm64 -gcflags=-S); derive mdctScale
= 1/9 = 2/N (N=18) properly instead of comparing a gain against a
normalization factor as if commensurate; stop overclaiming that Reset is
needed for the cross-package test and add TestFilterbankReset closing
its coverage; add an out-of-range guard and panic to bits.Writer.WriteBits
plus TestWriterPanicsOnBadN; wrap test-side floating products that feed
an adjacent accumulator with explicit float64() for arm64/amd64
determinism uniformity; and fix two stale (fbtables_test.go) comment
references to the file that actually holds those tests
(filterbank_test.go).

+++ internal/dec/encx_filterbank_test.go
@@ -50,7 +50,7 @@ func TestReconstructionGate(t *testing.T) {
 	x := make([]float64, totalSamples)
 	for n := range x {
 		t := float64(n) / sr
-		x[n] = a1*math.Sin(2*math.Pi*f1*t) + a2*math.Sin(2*math.Pi*f2*t)
+		x[n] = float64(a1*math.Sin(2*math.Pi*f1*t)) + float64(a2*math.Sin(2*math.Pi*f2*t))
 	}
 
 	pcm := make([]float32, totalSamples)
@@ -162,7 +162,7 @@ func measureDelay(x []float64, y []float32) int {
 			if j < 0 || j >= len(y) {
 				continue
 			}
-			sum += x[i] * float64(y[j])
+			sum += float64(x[i] * float64(y[j]))
 		}
 		if sum > bestCorr {
 			bestCorr = sum
@@ -184,8 +184,8 @@ func granuleRMSRatio(x []float64, y []float32, delay, start, length int) (gain f
 		if yi < 0 || yi >= len(y) {
 			continue
 		}
-		sumSqX += x[xi] * x[xi]
-		sumSqY += float64(y[yi]) * float64(y[yi])
+		sumSqX += float64(x[xi] * x[xi])
+		sumSqY += float64(float64(y[yi]) * float64(y[yi]))
 		n++
 	}
 	if n == 0 || sumSqX == 0 {
@@ -207,8 +207,8 @@ func computeSNR(x []float64, y []float32, delay int, gain float64, start, end in
 		}
 		recon := float64(y[j]) / gain
 		err := x[i] - recon
-		sigPower += x[i] * x[i]
-		noisePower += err * err
+		sigPower += float64(x[i] * x[i])
+		noisePower += float64(err * err)
 		n++
 	}
 	if n == 0 || noisePower == 0 {
